requirements.txt
inky
pillow
streamlit